+++ instapy/browser.py
@@ -4,6 +4,7 @@ from selenium.common.exceptions import WebDriverException
 from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.firefox.options import Options as Firefox_Options
 from selenium.webdriver import Remote
+from selenium.common.exceptions import UnexpectedAlertPresentException
 from webdriverdownloader import GeckoDriverDownloader
 from random_user_agent.user_agent import UserAgent
 
@@ -61,6 +62,7 @@ def set_selenium_local_session(
     page_delay,
     geckodriver_path,
     browser_executable_path,
+    logfolder,
     logger,
 ):
     """Starts local session for a selenium server.
@@ -121,17 +123,21 @@ def set_selenium_local_session(
     # mute audio while watching stories
     firefox_profile.set_preference("media.volume_scale", "0.0")
 
-    # prevent  Hide Selenium Extension: error
+    # prevent Hide Selenium Extension: error
     firefox_profile.set_preference("dom.webdriver.enabled", False)
     firefox_profile.set_preference("useAutomationExtension", False)
     firefox_profile.set_preference("general.platform.override", "iPhone")
     firefox_profile.update_preferences()
+
+    # geckodriver log in specific user logfolder
+    geckodriver_log = "{}geckodriver.log".format(logfolder)
 
     # prefer user path before downloaded one
     driver_path = geckodriver_path or get_geckodriver()
     browser = webdriver.Firefox(
         firefox_profile=firefox_profile,
         executable_path=driver_path,
+        log_path=geckodriver_log,
         options=firefox_options,
     )
 
@@ -149,7 +155,15 @@ def set_selenium_local_session(
 
     sleep(1)
     # set mobile viewport (iPhone X)
-    browser.set_window_size(375, 812)
+    try:
+        browser.set_window_size(375, 812)
+    except UnexpectedAlertPresentException as exc:
+        logger.exception(
+            "Unexpected alert on resizing web browser!\n\t"
+            "{}".format(str(exc).encode("utf-8"))
+        )
+        close_browser(browser, False, logger)
+        return browser, "Unexpected alert on browser resize"
 
     message = "Session started!"
     highlight_print("browser", message, "initialization", "info", logger)